% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allom_sub.R
\docType{data}
\name{allom.sub}
\alias{allom.sub}
\title{Mass-height data for Colt State Park, Rhode Island, USA}
\format{dataframe}
\usage{
data(allom.sub)
}
\description{
Mass-height data for Colt State Park, Rhode Island, USA
}
\keyword{allometry}
\keyword{datasets,}
