% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat1.R
\docType{data}
\name{dat1}
\alias{dat1}
\title{Stem height data for each mesocosm and each measurement date}
\format{dataframe}
\usage{
data(dat1)
}
\description{
Stem height data for each mesocosm and each measurement date
}
\keyword{NAPP}
\keyword{aboveground}
\keyword{biomass,}
\keyword{datasets,}
