% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\docType{data}
\name{dea}
\alias{dea}
\title{Denitrification enzyme activity and in vitro N2O production rates}
\format{dataframe}
\usage{
data(dea)
}
\description{
Denitrification enzyme activity and in vitro N2O production rates
}
\keyword{N2O}
\keyword{datasets,}
\keyword{denitrification,}
