% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lam.R
\name{lam}
\alias{lam}
\title{Identify lambda values from Box Cox model (based on Lu et al. 2016)}
\usage{
lam(df, mass = "sample", height = "height_cm", lam.avail = c(-2, -1.5, -1,
  -2/3, -1/2, -1/3, 0, 1/3, 1/2, 2/3, 1, 1.5, 2))
}
\arguments{
\item{df}{data frame}

\item{mass}{mass column}

\item{height}{height column}

\item{lam.avail}{possible lambda values that can be used in Box-Cox transformations}
}
\value{
lambda value
}
\description{
Identify lambda values from Box Cox model (based on Lu et al. 2016)
}
\examples{
\dontrun{}
}
