% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nappCalc2.R
\name{nappCalc2}
\alias{nappCalc2}
\title{Calculate net aboveground primary production}
\usage{
nappCalc2(dataset, liveCol = "mass", deadCol = "dead", yearCol = "year",
  siteCol = "pot2", MilnerHughes = "TRUE", EOS = "FALSE",
  EOS_window = 1, summarize = "TRUE", timeCol = "day",
  annualReset = "TRUE")
}
\arguments{
\item{dataset}{data}

\item{liveCol}{live biomass}

\item{deadCol}{dead biomass}

\item{yearCol}{year}

\item{siteCol}{site/plot/experimental unit identifier}

\item{MilnerHughes}{Do you want Milner Hughes calculated?}

\item{EOS}{Do you want end-of-season-live calculated?}

\item{EOS_window}{window for EOSL}

\item{summarize}{should data summary be added (output will be a list)}

\item{timeCol}{time column (sequential measurements within each year)}

\item{annualReset}{should data be reset to zero each year}
}
\value{
list
}
\description{
Calculate net aboveground primary production
}
\examples{
\dontrun{}
}
