% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Calculates standard error}
\usage{
se(x)
}
\arguments{
\item{x}{numeric or integer}
}
\value{
value
}
\description{
Calculates standard error
}
\examples{
se(CN_mass_data$n_pct)
plyr::ddply(CN_mass_data, plyr::.(species, pool_label), plyr::summarise, se = se(n_pct))
}
