% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test2.R
\docType{data}
\name{test2}
\alias{test2}
\title{Nutrient, stable isotope, and mass data from destructive mesocosm harvests}
\format{dataframe}
\usage{
data(test2)
}
\description{
Nutrient, stable isotope, and mass data from destructive mesocosm harvests
}
\keyword{15N}
\keyword{datasets,}
\keyword{isotopes,}
\keyword{stable}
