% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nappCalc2.R
\name{nappCalc2}
\alias{nappCalc2}
\title{Calculate net aboveground primary production}
\usage{
nappCalc2(dataset, liveCol = "mass", deadCol = "dead",
  yearCol = "year", siteCol = "pot2", timeCol = "day",
  annualReset = "TRUE", MilnerHughes = "TRUE", EOS = "FALSE",
  EOS_window = 1, summarize = "TRUE")
}
\arguments{
\item{dataset}{data}

\item{liveCol}{live biomass}

\item{deadCol}{dead biomass}

\item{yearCol}{year}

\item{siteCol}{site/plot/experimental unit identifier}

\item{timeCol}{time column (sequential measurements within each year)}

\item{annualReset}{should data be reset to zero each year}

\item{MilnerHughes}{If "TRUE", Milner-Hughes NAPP is calculated}

\item{EOS}{If "TRUE", end-of-season live biomass is reported}

\item{EOS_window}{window for EOSL}

\item{summarize}{If "TRUE", output will be a list with two elements: incremental data and summary data}
}
\value{
list
}
\description{
Calculate net aboveground primary production
}
