% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmolHr_mgDay.R
\name{nmolHr_mgDay}
\alias{nmolHr_mgDay}
\title{Convert N2O units from nanomoles of N2O per hour to milligrams of N per day}
\usage{
nmolHr_mgDay(x)
}
\arguments{
\item{x}{numeric or integer value(s)}
}
\value{
numeric value
}
\description{
Convert N2O units from nanomoles of N2O per hour to milligrams of N per day
}
\examples{
nmolHr_mgDay(dea$DEA)
}
